Reverted to unsigned 32 bit, only modulo must be signed

+++ src/ntt.cpp
@@ -1,5 +1,5 @@
 #include <cmath>		/* log2(), pow() */
-#include <cstdint>		/* int32_t */
+#include <cstdint>		/* int32_t, uint32_t */
 #include <cstdlib> 		/* calloc() */
 
 #include "../include/utils.h"	/* modExp() */
@@ -15,16 +15,16 @@
  * @param prim	The primitive root as applied to length and modulus
  * @return 	The transformed vector
  */
-int32_t *naiveNTT(int32_t *vec, int32_t n, int32_t m, int32_t prim){
+uint32_t *naiveNTT(uint32_t *vec, uint32_t n, uint32_t m, uint32_t prim){
 
-	int32_t *result;
-	result = (int32_t *) malloc(n*sizeof(int32_t));
+	uint32_t *result;
+	result = (uint32_t *) malloc(n*sizeof(uint32_t));
 
-	int32_t temp;
-	for(int32_t i = 0; i < n; i++){
+	uint32_t temp;
+	for(uint32_t i = 0; i < n; i++){
 
 		temp = 0;
-		for(int32_t j = 0; j < n; j++){
+		for(uint32_t j = 0; j < n; j++){
 
 			temp = temp + modulo(vec[j]*modExp(prim, i*j, m),m);
 
@@ -47,7 +47,7 @@ int32_t *naiveNTT(int32_t *vec, int32_t n, int32_t m, int32_t prim){
  * @param prim	The primitive root as applied to length and modulus
  * @return 	The transformed vector
  */
-int32_t *outOfPlaceNTT(int32_t *vec, int32_t n, int32_t m, int32_t prim){
+uint32_t *outOfPlaceNTT(uint32_t *vec, uint32_t n, uint32_t m, uint32_t prim){
 
 	if(n == 1){
 		
@@ -55,31 +55,31 @@ int32_t *outOfPlaceNTT(int32_t *vec, int32_t n, int32_t m, int32_t prim){
 
 	}
 
-	int32_t halfN = n >> 1;
+	uint32_t halfN = n >> 1;
 
-	int32_t *A0;
-	A0 = (int32_t *) malloc(halfN * sizeof(int32_t));
+	uint32_t *A0;
+	A0 = (uint32_t *) malloc(halfN * sizeof(uint32_t));
 
-	int32_t *A1;
-	A1 = (int32_t *) malloc(halfN * sizeof(int32_t));
+	uint32_t *A1;
+	A1 = (uint32_t *) malloc(halfN * sizeof(uint32_t));
 
-	for(int32_t i = 0; i < halfN; i++){
+	for(uint32_t i = 0; i < halfN; i++){
 
 		A0[i] = vec[i*2];
 		A1[i] = vec[i*2 + 1];
 
 	}
 
-	int32_t *y0 = outOfPlaceNTT(A0, halfN, m, prim);
-	int32_t *y1 = outOfPlaceNTT(A1, halfN, m, prim);
+	uint32_t *y0 = outOfPlaceNTT(A0, halfN, m, prim);
+	uint32_t *y1 = outOfPlaceNTT(A1, halfN, m, prim);
 
-	int32_t *result;
-	result = (int32_t *) calloc(n, sizeof(int32_t));
+	uint32_t *result;
+	result = (uint32_t *) calloc(n, sizeof(uint32_t));
 
-	for(int32_t i = 0; i < halfN; i++){
+	for(uint32_t i = 0; i < halfN; i++){
 
-		result[i] 		= modulo(y0[i] + modulo(modExp(prim,i,m)*y1[i],m),m);
-		result[i + halfN] 	= modulo(y0[i] - modulo(modExp(prim,i,m)*y1[i],m),m);
+		result[i] 		= modulo((int32_t)y0[i] + (int32_t)modulo(modExp(prim,i,m)*y1[i],m),m);
+		result[i + halfN] 	= modulo((int32_t)y0[i] - (int32_t)modulo(modExp(prim,i,m)*y1[i],m),m);
 
 	}
 
@@ -87,10 +87,10 @@ int32_t *outOfPlaceNTT(int32_t *vec, int32_t n, int32_t m, int32_t prim){
 
 }
 
-int32_t *inPlaceNTT(int32_t *vec, int32_t n){
+uint32_t *inPlaceNTT(uint32_t *vec, uint32_t n){
 
-	int32_t *result;
-	result = (int32_t *) calloc(n, sizeof(int32_t));
+	uint32_t *result;
+	result = (uint32_t *) calloc(n, sizeof(uint32_t));
 
 	return result;
